requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.8.2
email-validator==1.1.3
python-multipart==0.0.5
aiofiles==0.7.0
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
joblib==1.0.1 